Improvements to decoration plugin (#1214)
* Improvements to decoration plugin

* Button size bugfix

* Minor adjustments to button colors

* Light code cleanup; Add suitable comments

* Uncrustify plugins/decor/deco-theme.cpp

+++ plugins/decor/deco-theme.hpp
@@ -41,11 +41,11 @@ class decoration_theme_t
     struct button_state_t
     {
         /** Button width */
-        int width;
+        double width;
         /** Button height */
-        int height;
+        double height;
         /** Button outline size */
-        int border;
+        double border;
         /** Progress of button hover, in range [-1, 1].
          * Negative numbers are usually used for pressed state. */
         double hover_progress;
